requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==3.0.2
django-crispy-forms==1.8.1
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
